Update myscheduler.py
Changes the behavior for the following problem:

An example:
Shutter A is fully opened in the morning an closed for 60% by scheduled operations lets say at 8:00. Shutter A should again be closed by scheduled operations for 60% at 13:00. Between 8:00 and 13:00 the shutter is operated manually.

What happens?

Lets say Shutter A is moved manually to position 40% between 8:00 and 12:00 --> At 13:00 nothing happens. You will find an error in the logfile

Lets say Shutter A is moved manually to position 80% between 8:00 and 12:00 --> At 13:00 Shutter A is moved to positon 60%. Everything is fine

Lets say Shutter A is not moved manually between 8:00 and 12:00 at all --> At 13:00 Shutter A is fully closed.